Update map tab map to use mapRefState so we can watch for changes
