+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/useMapRef.js
@@ -1,0 +1,17 @@
+import { useCallback, useState } from "react";
+
+/**
+ * Use a callback ref to get the map instance and store it in state so we can watch it with useEffect
+ * @see https://reactjs.org/docs/refs-and-the-dom.html#callback-refs
+ * @returns {Array} - [mapRef, mapRefState] - mapRef is a callback ref, mapRefState is a state variable
+ */
+export const useMapRef = () => {
+  const [mapRefState, setMapRefState] = useState(null);
+  const mapRef = useCallback((mapInstance) => {
+    if (mapInstance !== null) {
+      // Store instance as the value of current just like a ref would
+      setMapRefState({ current: mapInstance });
+    }
+  }, []);
+  return [mapRef, mapRefState];
+};
